requirements.txt
dlib==19.24.0
numpy
opencv_python_headless==4.6.0.66
Pillow==9.4.0
streamlit==1.18.1
